Return the registry sort rank as a float so sub-ranks survive
get_attribute_order cast the CSV's float rank to int, so the fractional sub-ranks (20.1-20.3, 21.1-21.2) collapsed onto their parent and the intended intra-group order of curated columns fell back to incoming column order.

+++ src/openplaces/io/curator/formatters.py
@@ -127,7 +127,10 @@ _BARE_ID = re.compile(r'^(.+?)_id(_.+)?$')
 _BIG = 10_000
 
 
-def _attr_rank(name: str) -> int:
+def _attr_rank(name: str) -> float:
+    # Float, not int: the registry's fractional sub-ranks (20.1, 20.2)
+    # order attributes within one group, and rounding them to their
+    # parent rank loses that order.
     from openplaces.core.attribute_registry import get_attribute_order
 
     rank = get_attribute_order(resolve_attribute_name(name))
